Retire the Approvals screen: Notifications is the approvals surface
The notification expansion now carries everything the standalone screen
had, so the screen goes. What moved rather than died:

- Decide actions: new shared ApprovalBatchActions (approval-batch.tsx) —
  the retired screen's respond mutation verbatim (approve-all behind the
  Face ID signature, reject-all with an optional typed reason; approvals
  always sign, rejections never do). A still-open, unexpired batch renders
  them inside its notification expansion; onDecided refetches the batch.
- Approver key UI: new ApproverKeyBanner component (enroll / re-enroll /
  signing-as, moved verbatim) mounted above the Notifications list. Both
  it and the actions read the same approver-key-status query cache.
- Routing: approvals-destination pushes (scope holds still emit these)
  land on the Notifications view, which pre-expands the matching row
  (ActivityCard's toggled-or-default pattern). No scroll-to: the list is
  newest-first, so a fresh push's row is at or near the top.
- Gone: approvals.tsx, the drawer entry, the expansion's now-moot "Open in
  Approvals" link, and focusOpenBatch (screen-only helper) with its test.
